pyproject.toml
[build-system]
requires = ["setuptools<=64.0.3", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "readpy_P-Y-R-O-B-O-T"
version = "1.0.2"
description = "A MINI AND EASY TO USE REDIS SERVER"
readme = "README.md"
authors = [{name =  "KUSHAL", email = "NONE@NONE.com"}]
license = {file = "LICENSE"}
requires-python = ">=3.10"
keywords = ["redis", "memory", "communication",
			"RAM", "RAM server", "memory server",
			"shared memory", "network memory",
			"shared ram", "networked memory",
			"distributed memory", "multithreaded",
			"distributed RAM"
			]
dependencies = ["cypher-protocol-P-Y-R-O-B-O-T"]

[project.urls]
Homepage = "https://github.com/P-Y-R-O-B-O-T/READPY"
